- latestThreads -> isset - h3 in h2 umgewandelt

+++ Vpc/Forum/LatestThreads/Simple/Component.tpl
@@ -1,13 +1,15 @@
-<div class="<?=$this->cssClass ?>">
-    <h3>Aktuell im Forum:</h3>
-    <ul>
-        <?php $x= 0; foreach ($this->threads as $thread) { $x++ ?>
-            <li <?php if($x == 1) { ?> class="first"<?php } ?>>
-                <div class="thread">
-                    <?= $this->date($thread->lastPost->row->create_time) ?><br />
-                    <?=$this->componentLink($thread, $this->truncate($thread->row->subject, 50, '...', true));?>
-                </div>
-            </li>
-        <?php } ?>
-    </ul>
-</div>
+<? if ($this->threads) { ?>
+    <div class="<?=$this->cssClass ?>">
+        <h2>Aktuell im Forum:</h2>
+        <ul>
+            <?php $x= 0; foreach ($this->threads as $thread) { $x++ ?>
+                <li <?php if($x == 1) { ?> class="first"<?php } ?>>
+                    <div class="thread">
+                        <?= $this->date($thread->lastPost->row->create_time) ?><br />
+                        <?=$this->componentLink($thread, $this->truncate($thread->row->subject, 50, '...', true));?>
+                    </div>
+                </li>
+            <?php } ?>
+        </ul>
+    </div>
+<? } ?>
